+++ CHANGELOG.md
@@ -4,4 +4,7 @@ All notable changes to `crumbform` will be documented in this file
 
 ## 1.0.0 - XXXX-XX-XX
 
-- initial release
+- Logging: Database and file
+- Form validation
+- Multiple contact forms
+- Example with and without javascript